Fix styles

+++ src/controller/UserController.ts
@@ -1,27 +1,26 @@
-import {getRepository} from "typeorm";
-import {NextFunction, Request, Response} from "express";
-import {User} from "../entity/User";
+import { getRepository } from "typeorm";
+import { NextFunction, Request, Response } from "express";
+import { User } from "../entity/User";
 
 export class UserController {
-    private userRepository = getRepository(User);
+  private userRepository = getRepository(User);
 
-    async all(request: Request, response: Response) {
-        return this.userRepository.find();
-    }
+  async all(request: Request, response: Response) {
+    return this.userRepository.find();
+  }
 
-    // http://localhost:8080/user/61cb8827a45ace8fcf0ec1c1
-    async one(request: Request, response: Response) {
-        console.log(request.params.id)
-        return this.userRepository.findOne(request.params.id);
-    }
+  // http://localhost:8080/user/61cb8827a45ace8fcf0ec1c1
+  async one(request: Request, response: Response) {
+    console.log(request.params.id);
+    return this.userRepository.findOne(request.params.id);
+  }
 
-    async save(request: Request, response: Response) {
-        return this.userRepository.save(request.body);
-    }
+  async save(request: Request, response: Response) {
+    return this.userRepository.save(request.body);
+  }
 
-    async remove(request: Request, response: Response, next: NextFunction) {
-        let userToRemove = await this.userRepository.findOne(request.params.id);
-        if (userToRemove)
-          await this.userRepository.remove(userToRemove);
-    }
+  async remove(request: Request, response: Response, next: NextFunction) {
+    let userToRemove = await this.userRepository.findOne(request.params.id);
+    if (userToRemove) await this.userRepository.remove(userToRemove);
+  }
 }
